When a website has a big category tree, without this patch it has to make a lot of queries just to build the tree. Now it's way faster because it solves all in 1 query.
For example, in a production database, and after warmup, this reduced the amount of queries from 369 to 229 while browsing a category. That's 140 less queries, a lot. Also, when browsing /shop, it reduced the queries from 326 to 172. The times inside and outside Postgres also improve according to my tests.

Note: this domain could be even more constrained to avoid reading all website categories if they are collapsed, but given faed8d5 is not in v12, it ends up producing lots of more queries (by using child_of or parent_of operators) and slowing more the result than just downloading all of them

https://github.com/odoo/odoo/pull/60834